Split out parse eTag
